requirements.txt
openai
streamlit
requests
Pillow